requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
boto3>=1.34.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0

# MongoDB
motor>=3.3.0
pymongo>=4.6.0

# Visualization (optional - for testing)
opencv-python>=4.8.0
numpy>=1.26.0

# Object Detection Backends (install only what you need)
roboflow>=1.1.0
